Move exclusively() into util module.
